Re-organize the structure of DEPLOYSPEC; remove operations.py
This fulfills the idea of inferring the operations needed on a given
inventory based on the inventory kind and api_version. I don't know
why I didn't notice this before but the new structure of the DEPLOYSPEC
dir is very similar to how I structured the Ansible equivalent. Just
declarative yamls and no imperative list of operations.